#4 Implement test provider in git
