+++ tracklater/timemodules/gitmodule.py
@@ -2,16 +2,23 @@ from typing import List
 import settings
 import git
 import pytz
+import os
+import json
 
 from datetime import datetime
-from utils import FixedOffset
-from timemodules.interfaces import EntryMixin, AbstractParser, Entry
+from utils import FixedOffset, obj_from_dict
+from timemodules.interfaces import EntryMixin, AbstractParser, Entry, AbstractProvider
+
+
+import logging
+logger = logging.getLogger(__name__)
 
 
 def get_setting(key, default=None, group='global'):
     return settings.helper('GIT', key, group=group, default=default)
 
 
+FIXTURE_DIR = os.path.dirname(os.path.realpath(__file__)) + "/fixture"
 HEL = pytz.timezone('Europe/Helsinki')
 
 
@@ -26,28 +33,50 @@ class Parser(EntryMixin, AbstractParser):
         start_date = self.start_date.replace(tzinfo=HEL)
         end_date = self.end_date.replace(tzinfo=HEL)
         log = []
+        provider = Provider()
         for group, data in settings.GIT.items():
             for repo_path in data.get('REPOS', []):
-                repo = git.Repo(repo_path)
-                for branch in repo.branches:
-                    try:
-                        entries = branch.log()
-                    except Exception:
+                for log_entry in provider.get_log_entries(repo_path):
+                    if log_entry.actor.email not in settings.GIT['global']['EMAILS']:
                         continue
-                    for log_entry in entries:
-                        if log_entry.actor.email not in settings.GIT['global']['EMAILS']:
-                            continue
-                        if not log_entry.message.startswith('commit'):
-                            continue
-                        message = ''.join(log_entry.message.split(':')[1:])
-                        time = timestamp_to_datetime(log_entry.time)
-                        if time < start_date or time > end_date:
-                            continue
+                    if not log_entry.message.startswith('commit'):
+                        continue
+                    message = ''.join(log_entry.message.split(':')[1:])
+                    time = timestamp_to_datetime(log_entry.time)
+                    if time < start_date or time > end_date:
+                        continue
 
-                        log.append(Entry(
-                            text=[
-                                "{} - {}".format(repo_path.split('/')[-1], message)
-                            ],
-                            start_time=time,
-                        ))
+                    log.append(Entry(
+                        text=[
+                            "{} - {}".format(repo_path.split('/')[-1], message)
+                        ],
+                        start_time=time,
+                    ))
         return log
+
+
+class Provider(AbstractProvider):
+    def get_log_entries(self, repo_path):
+        repo = git.Repo(repo_path)
+        for branch in repo.branches:
+            try:
+                entries = branch.log()
+            except Exception as e:
+                logger.warning(e)
+                continue
+            for log_entry in entries:
+                yield log_entry
+
+    def test_get_log_entries(self, repo_path):
+        with open(FIXTURE_DIR + '/git_test_data.json', 'r') as f:
+            _git = obj_from_dict(json.load(f))
+
+        repo = _git.Repo(repo_path)
+        for branch in repo.branches:
+            try:
+                entries = branch.log()
+            except Exception as e:
+                logger.warning(e)
+                continue
+            for log_entry in entries:
+                yield log_entry
